Enum zero_bin_common::block_interval::BlockInterval

pub enum BlockInterval {
    SingleBlockId(BlockId),
    Range(Range<u64>),
    FollowFrom {
        start_block: u64,
        block_time: Option<u64>,
    },
}

Range of blocks to be processed and proven.

Variants

SingleBlockId(BlockId)

Range(Range<u64>)

FollowFrom

Fields

start_block: u64

block_time: Option<u64>

Implementations

impl BlockInterval

pub fn new(s: &str) -> Result<BlockInterval>

Create a new block interval

A valid block range is of the form: * block_number for a single block number * lhs..rhs, lhs..=rhs as an exclusive/inclusive range * lhs.. for a range starting from lhs to the chain tip. lhs..= is also valid format.

Example

   assert_eq!(BlockInterval::new("0..10").unwrap(), BlockInterval::Range(0..10));
   assert_eq!(BlockInterval::new("0..=10").unwrap(), BlockInterval::Range(0..11));
   assert_eq!(BlockInterval::new("32141").unwrap(), BlockInterval::SingleBlockId(BlockId::Number(32141.into())));
   assert_eq!(BlockInterval::new("100..").unwrap(), BlockInterval::FollowFrom{start_block: 100, block_time: None});

pub fn into_bounded_stream(self) -> Result<impl Stream<Item = u64>>

Convert the block interval into an async stream of block numbers.

pub fn get_start_block(&self) -> Result<u64>

pub async fn into_unbounded_stream<ProviderT, TransportT>( self, provider: ProviderT, ) -> Result<impl Stream<Item = Result<u64, Error>>, Error>
where ProviderT: Provider<TransportT>, TransportT: Transport + Clone,

Convert the block interval into an unbounded async stream of block numbers. Query the blockchain node for the latest block number.