Module zero_bin_common::prover_state

Global prover state management and utilities.

This module provides the following:

[Circuit] and CircuitConfig which can be used to dynamically construct [evm_arithmetization::fixed_recursive_verifier::AllRecursiveCircuits] from the specified circuit sizes.

Command line arguments for constructing a CircuitConfig.

Provides default values for the circuit sizes.

Allows the circuit sizes to be specified via environment variables.

Persistence utilities for saving and loading [evm_arithmetization::fixed_recursive_verifier::AllRecursiveCircuits].

Global prover state management via the [P_STATE] static and the [set_prover_state_from_config] function.

Modules

circuit

[AllRecursiveCircuits] dynamic circuit configuration.

cli

CLI arguments for constructing a CircuitConfig, which can be used to construct table circuits.

persistence

Structs

ProverStateManager

Product of CircuitConfig and CircuitPersistence.

Enums

CircuitPersistence

Specifies whether to persist the processed circuits.

TableLoadStrategy

Specifies how to load the table circuits.

Functions

p_manager

p_state