Module zero_bin_common::prover_state::persistence

Statics

CIRCUIT_VERSION

When we serialize/deserialize circuits, we rely on the hash of the plonky kernel to determine if the circuit is compatible with our current binary. If the kernel hash of the circuit that we are loading in from disk differs, then using these circuits would cause failures during proof generation

Functions

delete_all

Flushes all existing prover state configurations and associated circuits that have been written to disk.

persist_all_to_disk

Writes the provided [AllRecursiveCircuits] to disk with all configurations, along with the associated [VerifierData].

set_circuit_cache_dir_env_if_not_set

We store serialized circuits inside the cache directory specified by an env variable. If the user does not set this, then we set it base to the OS’s standard location for the cache directory.